+++ terraform/environments/staging/main.tf
@@ -1,9 +1,10 @@
 terraform {
-  backend "s3" {
-    bucket         = "nn-terraform-state-eu"
-    key            = "staging/terraform.tfstate"
-    region         = "eu-central-1"
-    dynamodb_table = "nn-terraform-locks"
+  cloud {
+    organization = "notablenomads"
+
+    workspaces {
+      name = "nn-backend-api-staging"
+    }
   }
 
   required_providers {
@@ -23,17 +24,18 @@ module "vpc" {
 
   app_name    = var.app_name
   environment = var.environment
+  vpc_cidr    = var.vpc_cidr
 }
 
 # Get the zone ID from the shared state
 data "terraform_remote_state" "shared" {
-  backend = "s3"
+  backend = "remote"
   
   config = {
-    bucket         = "nn-terraform-state-eu"
-    key            = "shared/terraform.tfstate"
-    region         = "eu-central-1"
-    dynamodb_table = "nn-terraform-locks"
+    organization = "notablenomads"
+    workspaces = {
+      name = "nn-backend-api-shared"
+    }
   }
 }
 
@@ -47,7 +49,7 @@ resource "aws_ssm_parameter" "env_variables" {
     API_PREFIX          = "v1"
     CORS_ENABLED_DOMAINS = "*.notablenomads.com,notablenomads.com"
     CORS_RESTRICT       = "false"
-    LOG_LEVEL           = "error"
+    LOG_LEVEL           = "debug"
     EMAIL_FROM_ADDRESS  = "no-reply@mail.notablenomads.com"
     EMAIL_TO_ADDRESS    = "contact@notablenomads.com"
     ***REMOVED***          = var.aws_region
@@ -56,7 +58,6 @@ resource "aws_ssm_parameter" "env_variables" {
   name      = "/platform/staging/${each.key}"
   type      = "String"
   value     = each.value
-  overwrite = true
   tags = {
     Environment = var.environment
     Type        = "Environment Variable"
@@ -73,7 +74,7 @@ resource "aws_ssm_parameter" "env_variables" {
 
 # You can choose either EC2 or ECS deployment by commenting/uncommenting the appropriate module
 module "api" {
-  source = "../../modules/api_ec2"  # For EC2 deployment
+  source = "../../modules/api_ec2"
 
   app_name            = var.app_name
   environment         = var.environment
@@ -86,11 +87,10 @@ module "api" {
   desired_count       = 1
   log_retention_days  = 1
   ssm_prefix         = "/platform/staging"
-  domain_name        = "api.staging.platform.notablenomads.com"
+  domain_name        = var.domain_name
   zone_id            = data.terraform_remote_state.shared.outputs.platform_zone_id
+  instance_type      = "t4g.nano"
   
-  # We don't need to specify environment_variables and secrets here anymore
-  # as they are managed by SSM parameters above
   environment_variables = []
   secrets = []
 }