+++ terraform/modules/vpc/outputs.tf
@@ -1,6 +1,6 @@
 output "vpc_id" {
   description = "The ID of the VPC"
-  value       = aws_vpc.this[0].id
+  value       = aws_vpc.main.id
 }
 
 output "public_subnet_ids" {
@@ -11,4 +11,14 @@ output "public_subnet_ids" {
 output "private_subnet_ids" {
   description = "List of private subnet IDs"
   value       = aws_subnet.private[*].id
+}
+
+output "public_route_table_id" {
+  description = "ID of the public route table"
+  value       = aws_route_table.public.id
+}
+
+output "private_route_table_id" {
+  description = "ID of the private route table"
+  value       = aws_route_table.private.id
 }